FileAPI/BlobURL/Blob-File-URL-Test2.html

<!DOCTYPE html>
<html>
<body>
<head>
<meta charset=utf-8>
<title>Blob and File reference URL Test(2)</title>
<link rel=help href="http://www.w3.org/TR/FileAPI/#convenienceAPI">
<link rel=author title="Breezewish" href="mailto:me@breeswish.org">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
</head>
<body>
    <form name="upload">
        <input type="file" id="fileChooser"><br><input type="button" id="start" value="start">
    </form>

    <div>
        <p>Test steps:</p>
        <ol>
            <li>Download the <a href="support/file_test1.js">file</a>.</li>
            <li>Select the file in the file inputbox.</li>
            <li>Delete the file.</li>
            <li>Click the 'start' button.</li>
        </ol>
    </div>

    <div id="log"></div>

    <script>

        var fileChooser = document.querySelector('#fileChooser');

        setup({explicit_done: true});
        setup({explicit_timeout: true});

        on_event(document.querySelector('#start'), 'click', function() {

            (function() {

                var t = async_test('Check whether the browser response 500 in XHR if the selected file which File/Blob URL refered is not found');
                t.step(function() {


                    var url = URL.createObjectURL(fileChooser.files[0]);

                    var xhr = new XMLHttpRequest();
                    xhr.open('GET', url, true);
                    xhr.onreadystatechange = function(e) {

                        if (this.readyState == 4) {

                            t.step_func_done(function() {

                                assert_equals(this.status, 500, 'status code should be 500.');
                                t.done();

                            }, this)();

                        }
                    }
                    xhr.send();

                });

            })();


            done();

        });
        

    </script>
</body>
</html>
